+++ celements-xwiki-core/src/main/java/com/celements/store/id/IdVersion.java
@@ -1,4 +1,4 @@
-package com.xpn.xwiki.store;
+package com.celements.store.id;
 
 public enum IdVersion {
